updating config
